backend/requirements.txt
Flask
flask-cors
mediapipe
numpy
opencv-python
Pillow
